Test (and fix) freeing split list images
Adding Destroy() calls to the split test found a bug, which was that I was
assigning a pointer to the same ImageInfo location to each split tile. When
freed, the first Destroy() call would free the ImageInfo location, but
subsequent ones would be attempting to free already-freed memory, causing
panics.